Add military airports and routing

+++ src/Position.java
@@ -1,22 +1,20 @@
-import javafx.geometry.Pos;
-
 /**
  * Created by Tymek on 13.10.15.
  */
 public class Position {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public Position(){
 
     }
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
@@ -24,7 +22,7 @@ public class Position {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
